+++ run.php
@@ -30,7 +30,7 @@ function plugin(): Plugin {
 			'MIN_MYSQL_VERSION'            => '',
 			'PLUGIN_PREFIX'                => 'csbxwoo',
 			'PLUGIN_NAME'                  => 'BoxNow Delivery for WooCommerce by CodeSoup',
-			'PLUGIN_VERSION'               => '1.0.0',
+			'PLUGIN_VERSION'               => '1.1.0',
 			'PLUGIN_TEXTDOMAIN'            => 'codesoup-woo-boxnow',
 			'ENVIRONMENT'                  => \wp_get_environment_type(),
 		);